[FIX] Change class CarrierAccount in its own python file
