+++ base_delivery_carrier_label/__init__.py
@@ -20,3 +20,4 @@
 ##############################################################################
 from . import delivery
 from . import stock
+from . import carrier_account
